This is why I love scaffolding!
Since this project is a labor of love that scratches my itch I just continue to try to use the app as I want it to work. My first steps in creating Lettucebooks is to get the general workflow to go as I want it and have the features I need to manage my own finances. So far so good...